release: Run other workflows explicitly
To bypass restrictions on GitHub Actions run the docs and ci workflows
manually after creating a release. GitHub Actions does not trigger
other workflows from a workflow run if not using a personal access
token (PAT). By calling the workflows explicitly, it is possible to
work around this restriction without the need for a PAT.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>